speech-recognition-config-tests/requirements.txt
cdp-backend
gitpython